requirements.txt
streamlit
openai
python-dotenv
watchdog
PyPDF2
python-docx
python-pptx
matplotlib
pandas
